+++ Sources/F02CategoriesAndProtocols/NSString+PFStudyFormatting.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface NSString (PFStudyFormatting)
+- (NSString *)pf_titleCaseLabel;
+@end
+
+NS_ASSUME_NONNULL_END