+++ Sources/F02CategoriesAndProtocols/NSString+PFStudyFormatting.m
@@ -1,0 +1,9 @@
+#import "NSString+PFStudyFormatting.h"
+
+@implementation NSString (PFStudyFormatting)
+
+- (NSString *)pf_titleCaseLabel {
+    return [self capitalizedString];
+}
+
+@end